Actualización con correción de nombres de carpetas en mayusculas
